requirements.txt
langchain
langchain-community
langchain-openai
langchain-core
python-dotenv
requests 
faiss-cpu
streamlit 
ipykernel